backend/requirements.txt
# FastAPI Framework
fastapi==0.115.12
uvicorn[standard]==0.34.3

# Machine Learning & AI
torch>=2.0.0
transformers>=4.30.0
huggingface_hub>=0.20.0
tokenizers>=0.15.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.6.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Requests & Utilities
requests>=2.31.0
aiofiles>=23.0.0

# Email validation
email-validator>=2.1.0

# Google OAuth
google-auth>=2.25.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.23.0
